build,fix:Build fix for Windows by casting to int64_t because that is what is supported by ConfigurationElement::ValueType. - It is unclear why macOS does not need this.

+++ src/core/Group.cpp
@@ -21,7 +21,7 @@ namespace dagui
     {
         dagbase::ConfigurationElement::ValueType retval;
 
-        retval = dagbase::findEndpoint(path, "numChildren", std::uint32_t(_children.a.size()));
+        retval = dagbase::findEndpoint(path, "numChildren", std::int64_t(_children.a.size()));
         if (retval.has_value())
             return retval;
